==> notebooks/12-appendix-load-dataset-and-vectors.ipynb <==
# Load the dataset with vectors (E5-Large)


## 1. Load & Process


### 1.1 Load as DF and filter out by tokens threshold (40)


import pandas as pd
import os
import json
from tqdm import tqdm

df = pd.DataFrame(columns=["community", "user_name", "user_texts", "#_of_long_texts"])


def process_files(community, df):
    threshold = 40
    directory = f"../data/detailed_filtered_data/texts/{community}"
    for filename in tqdm(os.listdir(directory)):
        if filename.endswith(".json"):
            user_texts = []
            with open(os.path.join(directory, filename), "r") as f:
                data = json.load(f)
                for key in data:
                    for submission in data[key].get("submissions", []):
                        title = submission.get("title", "")
                        body = submission.get("body", "")
                        if len(title.split()) >= threshold:
                            user_texts.append(title)
                        if len(body.split()) >= threshold:
                            user_texts.append(body)
                    for submission in data[key].get("comments", []):
                        body = submission.get("body", "")
                        if len(body.split()) >= threshold:
                            user_texts.append(body)
            if len(user_texts) == 0:
                continue
            user_name = filename.split(".json")[0]
            new_row = pd.DataFrame(
                {
                    "community": community,
                    "user_name": [user_name],
                    "user_texts": [user_texts],
                    "#_of_long_texts": [len(user_texts)],
                }
            )
            df = pd.concat([df, new_row], ignore_index=True)
    return df


communities = [
    "classicalmusic",
    "electronicmusic",
    "hiphopheads",
    "indieheads",
    "Metal",
]

for vector_file_name in communities:
    df = pd.concat([df, process_files(vector_file_name, df)])
print("Done processing texts :)")

### 1.2 De-Duplication


def deduplicate_texts(df):
    exploded_df = df.explode("user_texts")
    exploded_df = exploded_df.drop_duplicates(subset="user_texts")
    return exploded_df


df = deduplicate_texts(df)
print("After dedeuplication there are", format(len(df), ","), "unique texts in total")

### 1.3 Split DF by Genre


classicalmusic_df = df[df["community"] == "classicalmusic"]
electronicmusic_df = df[df["community"] == "electronicmusic"]
hiphopheads_df = df[df["community"] == "hiphopheads"]
indieheads_df = df[df["community"] == "indieheads"]
metal_df = df[df["community"] == "Metal"]

communities_dfs = [
    classicalmusic_df,
    electronicmusic_df,
    hiphopheads_df,
    indieheads_df,
    metal_df,
]

print(
    "Number of texts left after deduplication per community:",
    "\nClassicalmusic:",
    format(len(classicalmusic_df), ","),
    "\nElectronicmusic:",
    format(len(electronicmusic_df), ","),
    "\nHiphopheads:",
    format(len(hiphopheads_df), ","),
    "\nIndieheads:",
    format(len(indieheads_df), ","),
    "\nMetal:",
    format(len(metal_df), ","),
)

### 1.4 Text-Count per user


classicalmusic_counts = classicalmusic_df["user_name"].value_counts()
electronicmusic_counts = electronicmusic_df["user_name"].value_counts()
hiphopheads_counts = hiphopheads_df["user_name"].value_counts()
indieheads_counts = indieheads_df["user_name"].value_counts()
metal_counts = metal_df["user_name"].value_counts()


print("Median and Mean of texts-count per user in each community after deduplication:")
print(
    "classicalmusic: Median =",
    classicalmusic_counts.median(),
    " | Mean =",
    round(classicalmusic_counts.mean(), 2),
)
print(
    "electronicmusic: Median =",
    electronicmusic_counts.median(),
    " | Mean =",
    round(electronicmusic_counts.mean(), 2),
)
print(
    "hiphopheads: Median =",
    hiphopheads_counts.median(),
    " | Mean =",
    round(hiphopheads_counts.mean(), 2),
)
print(
    "indieheads: Median =",
    indieheads_counts.median(),
    " | Mean =",
    round(indieheads_counts.mean(), 2),
)
print(
    "metal: Median =", metal_counts.median(), " | Mean =", round(metal_counts.mean(), 2)
)

## 2. Vectorize all Redditors texts


### 2.1 Vectorize


#### 2.1.1 Vectorization method


from sentence_transformers import SentenceTransformer
from typing import List, Union
from torch import Tensor
from numpy import ndarray


def vectorize_text(
    text, model: SentenceTransformer, batch_size=12, show_progress_bar=False
) -> Union[List[Tensor], ndarray, Tensor]:
    return model.encode(
        text, show_progress_bar=show_progress_bar, batch_size=batch_size
    )

#### 2.1.2 Vectorize & persist or reload (method only)


import h5py
import numpy as np

MODEL = "intfloat/e5-large-v2"

model = SentenceTransformer(MODEL)


def vectorize_or_load_community_data(community_name: str, df: pd.DataFrame):
    file_path = f"../data/detailed_filtered_data/vectors/{community_name}.h5"
    if os.path.exists(file_path):
        print(f"Loading vectors for {community_name} from file")
        with h5py.File(file_path, "r") as f:
            dataset = f["vectors"]
            df["vector"] = dataset[:].tolist()
    else:
        print(f"Vectorizing texts for {community_name}...")
        compression = "lzf"
        estimated_rows = len(df)

        with h5py.File(file_path, "a") as f:
            if "vectors" not in f:
                dataset = f.create_dataset(
                    "vectors",
                    (estimated_rows, model.get_sentence_embedding_dimension()),
                    dtype=np.float32,
                    compression=compression,
                )
                print("Dataset shape:", dataset.shape)
            else:
                dataset = f["vectors"]

            i = 0
            progress_bar = tqdm(total=estimated_rows, desc="Vectorizing texts")
            for _, row in df.iterrows():
                vector = vectorize_text(row["user_texts"], model)
                dataset[i] = vector
                i += 1
                progress_bar.update(1)
            progress_bar.close()

#### 2.1.3 Load vectors (or vectorize)


community_names = ["classical", "electronic", "hiphopheads", "indieheads", "metal"]
for community_name, df in zip(community_names, communities_dfs):
    vectorize_or_load_community_data(community_name, df)
df = pd.concat(communities_dfs, ignore_index=True)

### 2.4 Show Sample


df.head()

## 3. Predict personality traits presence


### 3.1 Load the traits models


import pickle
from sklearn.linear_model import LogisticRegression

traits_and_models = {"opn": None, "ext": None, "neu": None, "agr": None, "con": None}

for trait in traits_and_models.keys():
    with open(f"../models/step-10/gpt_{trait}.pkl", "rb") as f:
        traits_and_models[trait] = pickle.load(f)

### 3.2 Predict-Probablily per trait


def predict_trait_probab(trait: str, df: pd.DataFrame, model: LogisticRegression):
    df[f"{trait}_proba"] = df["vector"].apply(
        lambda x: model.predict_proba(np.array(x).reshape(1, -1))
    )
    return df

for trait, model in traits_and_models.items():
    df = predict_trait_probab(trait, df, model)

# That's it!

The dataset is loaded, along with the texts' vectors and the presence probability prediction per personality trait


# Print the first row of each community as example
df.groupby("community").head(1)

## How to vectorize (new) texts & predict 
presence of each personality traits in them? <br>
Just use the following code 👇

MODEL = "intfloat/e5-large-v2"


def embed_and_perdict_personality_trait_presence(texts: list[str]):
    """Text could be a single text or a list of texts."""
    model = SentenceTransformer(MODEL)
    embeddings = model.encode(texts, convert_to_tensor=True)
    result = {}
    for trait, model in traits_and_models.items():
        proba = model.predict_proba(embeddings)
        result.update({trait: proba})
    return result


texts = ["This is an example text 1"]
print(embed_and_perdict_personality_trait_presence(texts))

### Final experiment:
Predict music genre from personality probability vector <br>
The baseline is 0.2 (because we have 5 classes). So >0.2 beats the baseline.

df['personality_vector'] = df.apply(lambda row: [float(row['proba_for_high_opn']), float(row['proba_for_high_ext']), float(row['proba_for_high_neu']), float(row['proba_for_high_agr']), float(row['proba_for_high_con'])], axis=1)

from sklearn.linear_model import LogisticRegression
from sklearn.metrics import accuracy_score
from sklearn.model_selection import train_test_split

train_df, test_df = train_test_split(df, test_size=0.2, random_state=433)
model = LogisticRegression()
model.fit(train_df['personality_vector'].tolist(), train_df['community'])
community_prediction = model.predict(test_df['personality_vector'].tolist())
accuracy = accuracy_score(test_df['community'], community_prediction)
print(f"Accuracy: {accuracy}")

from sklearn.metrics import classification_report, confusion_matrix
import seaborn as sns

print("Classification Report:")
print(classification_report(test_df['community'], community_prediction, target_names=train_df['community'].unique()))
# Print Classification Report
classification_report = classification_report(test_df['community'], community_prediction, target_names=train_df['community'].unique(), output_dict=True)
classification_report_df = pd.DataFrame(classification_report).transpose()
classification_report_df = classification_report_df.drop(columns=['support'])

# Set a larger figure size
plt.figure(figsize=(10, 6))  # Adjust the figure size as needed

# Plot with a different colormap, e.g., 'viridis'
classification_report_df.plot(kind='bar', cmap='viridis')

plt.xlabel('Class')
plt.ylabel('Score')
plt.title('Classification Report')
plt.legend(loc='lower right')
plt.show()

conf_matrix = confusion_matrix(test_df['community'], community_prediction)
print("Confusion Matrix:")

class_labels = ['Metal', 'Classical', 'Electronic', 'Indie', 'HipHop']
plt.figure(figsize=(8, 6))
sns.heatmap(conf_matrix, annot=True, fmt="d", cmap="Blues", xticklabels=class_labels, yticklabels=class_labels)
plt.xlabel("Predicted")
plt.ylabel("Actual")
plt.title("Confusion Matrix")
plt.show()

from sklearn.model_selection import KFold, cross_val_score
from sklearn.linear_model import LogisticRegression
import numpy as np
from tqdm import tqdm
vals = []

X = np.array(df['personality_vector'].tolist())
y = np.array(df['community'])

kfold = KFold(n_splits=10, shuffle=True, random_state=307)
model = LogisticRegression()

accuracies = cross_val_score(model, X, y, cv=kfold, scoring='accuracy')

for i, accuracy in enumerate(accuracies, start=1):
    print(f"Fold {i}: Accuracy = {accuracy}")

print(f"Average Accuracy: {np.mean(accuracies)}")

print("Classification Report:")
print(classification_report(test_df['community'], community_prediction, target_names=train_df['community'].unique()))
# Print Classification Report
classification_report = classification_report(test_df['community'], community_prediction, target_names=train_df['community'].unique(), output_dict=True)
classification_report_df = pd.DataFrame(classification_report).transpose()
classification_report_df = classification_report_df.drop(columns=['support'])

# Set a larger figure size
plt.figure(figsize=(10, 6))  # Adjust the figure size as needed

# Plot with a different colormap, e.g., 'viridis'
classification_report_df.plot(kind='bar', cmap='viridis')

plt.xlabel('Class')
plt.ylabel('Score')
plt.title('Classification Report')
plt.legend(loc='lower right')
plt.show()

conf_matrix = confusion_matrix(test_df['community'], community_prediction)
print("Confusion Matrix:")

class_labels = ['Metal', 'Classical', 'Electronic', 'Indie', 'HipHop']
plt.figure(figsize=(8, 6))
sns.heatmap(conf_matrix, annot=True, fmt="d", cmap="Blues", xticklabels=class_labels, yticklabels=class_labels)
plt.xlabel("Predicted")
plt.ylabel("Actual")
plt.title("Confusion Matrix")
plt.show()
